refactor: 수정 및 삭제에 대한 로직 개선

+++ src/components/calender/ScheduleDrawer/layouts/ScheduleDrawerHeader.js
@@ -1,16 +1,20 @@
-import { Button, Stack, Typography } from '@mui/material';
+import {
+  Button, Stack, Tooltip, Typography,
+} from '@mui/material';
 import ClearIcon from '@mui/icons-material/Clear';
 import { useDispatch, useSelector } from 'react-redux';
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
 import { SCHEDULE_DRAWER, SCHEDULE_DRAWER_MODE } from '../../../../utils/constants/schedule';
 import { deleteSchedule, selectSchedule } from '../../../../utils/redux/schedule/scheduleSlice';
+import { selectGuestMode } from '../../../../utils/redux/common/commonSlice';
 
 function ScheduleDrawerHeader({ mode, setBottomDrawerOpen }) {
   const dispatch = useDispatch();
   const schedule = useSelector(selectSchedule);
+  const guestMode = useSelector(selectGuestMode);
 
   const deleteSelectedSchedule = () => {
-    if (window.confirm('정말로 삭제 하시겠습니까?')) {
+    if (guestMode && window.confirm('정말로 삭제 하시겠습니까?')) {
       dispatch(deleteSchedule(schedule.id));
       setBottomDrawerOpen(false);
     }
@@ -23,7 +27,20 @@ function ScheduleDrawerHeader({ mode, setBottomDrawerOpen }) {
 
       {
         mode === SCHEDULE_DRAWER_MODE.수정
-          ? <Button onClick={() => deleteSelectedSchedule()} color="error"><DeleteForeverIcon /></Button>
+          ? (
+            <Tooltip
+              title={!guestMode && '아직 일반 모드에서는 동작하지 않습니다.'}
+              placement="top"
+            >
+              <Button
+                onClick={() => deleteSelectedSchedule()}
+                color="error"
+              >
+                <DeleteForeverIcon />
+              </Button>
+            </Tooltip>
+
+          )
           : <Button disabled />
       }
     </Stack>
